splitting server up
