+++ main.go
@@ -4,19 +4,14 @@ import (
 	"log"
 	"net/http"
 
-	"github.com/gorilla/mux"
-	"github.com/nytlabs/st-core/core"
+	"github.com/nytlabs/st-core/core/server"
 )
 
 func main() {
 
-	s := core.NewServer()
-	r := mux.NewRouter()
-	r.HandleFunc("/", s.WebsocketHandler).Methods("GET")
-	r.HandleFunc("/group", s.GetGroupHandler).Methods("GET")
-	r.HandleFunc("/group", s.CreateGroupHandler).Methods("POST")
-	r.HandleFunc("/block", s.CreateBlockHandler).Methods("POST")
-	r.HandleFunc("/connection", s.CreateConnectionHandler).Methods("POST")
+	s := server.NewServer()
+	r := s.NewRouter()
+
 	http.Handle("/", r)
 
 	log.Println("serving on 7071")
